requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
clikit==0.6.2
cmdstanpy==1.3.0
comm==0.2.3
contourpy==1.3.3
crashtest==0.3.1
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
holidays==0.38
httpcore==1.0.9
httpstan==4.13.0
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
ipykernel==6.27.1
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.8.2
matplotlib-inline==0.2.1
mistune==3.2.0
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==1.26.2
openpyxl==3.1.2
packaging==26.0
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.6
pastel==0.2.1
patsy==1.0.2
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==5.18.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.1.5
psutil==7.2.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pylev==1.4.0
pyparsing==3.3.2
pysimdjson==7.0.2
pystan==3.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==4.0.0
pytz==2026.1.post1
PyYAML==6.0.1
pyzmq==27.1.0
qtconsole==5.7.1
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==2.1.0
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.23
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.1
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.5
tqdm==4.66.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webargs==8.7.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.46.3
widgetsnbextension==4.0.15
xlrd==1.2.0
yarl==1.23.0
